Wokring on camera controller
Changing the camera controller and the camera confiner. Not fully working on the wall right now. If need for testing change the gravity flip on the rift trigger and disable the new camera change the inspector camera on the rift too and enable the old camera and drag it in.